Importing loader from fms
