--- allocation/api/api.py
@@ -1,6 +1,0 @@
-from allocation.api.zyre import ZyreAPI
-
-
-class API(object):
-    def __init__(self, config):
-        pass
