+++ tests/data/non_overlapping_1.yaml
@@ -44,44 +44,4 @@ tasks:
     id: 73ace543-260e-474a-9112-3abd470ed5b5
     latest_start_time: 2776.44
     start_pose_name: shelf
-  988640de-f114-4b13-820d-7a410085ef96:
-    earliest_start_time: 2170.7
-    estimated_duration: 4.91
-    finish_pose_name: inspection_area
-    hard_constraints: true
-    id: 988640de-f114-4b13-820d-7a410085ef96
-    latest_start_time: 2303.49
-    start_pose_name: living_room
-  a4e10576-a0e3-4662-8aec-5992a75d2838:
-    earliest_start_time: 786.06
-    estimated_duration: 2.74
-    finish_pose_name: shelf
-    hard_constraints: true
-    id: a4e10576-a0e3-4662-8aec-5992a75d2838
-    latest_start_time: 1063.35
-    start_pose_name: cupboard
-  a71f1730-d5ed-45c5-9b3b-ad783869957c:
-    earliest_start_time: 2861.26
-    estimated_duration: 6.37
-    finish_pose_name: trash_can
-    hard_constraints: true
-    id: a71f1730-d5ed-45c5-9b3b-ad783869957c
-    latest_start_time: 3062.39
-    start_pose_name: shelf_cupboard
-  d36ba90a-7bee-4ee6-a7ce-82ac237457fb:
-    earliest_start_time: 1677.82
-    estimated_duration: 4.19
-    finish_pose_name: cupboard
-    hard_constraints: true
-    id: d36ba90a-7bee-4ee6-a7ce-82ac237457fb
-    latest_start_time: 1969.0
-    start_pose_name: observation_pose
-  fa1316ec-151c-46ea-8c62-72480b2654b5:
-    earliest_start_time: 1250.51
-    estimated_duration: 2.3
-    finish_pose_name: observation_pose
-    hard_constraints: true
-    id: fa1316ec-151c-46ea-8c62-72480b2654b5
-    latest_start_time: 1530.52
-    start_pose_name: living_room
 upper_bound: 300
